Week7_임윤섭_마이페이지에서 로그인된 유저 정보 받아오기
